Add userHandler to handle requests related to suer
