+++ src/middleware/authHandler.ts
@@ -1,5 +1,4 @@
 import { NavigateFunction } from 'react-router-dom';
-import { getUserRole } from './apiHandler';
 
 export const getAuthHeader = (token: string) => {
 	return { Authorization: `Bearer ${token}` };
@@ -16,22 +15,6 @@ export const setTokenToLocalStorage = async (data: Response) => {
 		}
 	} catch (error) {
 		console.error('There has been an issue while processing JSON: ', error);
-	}
-};
-
-export const getRole = async () => {
-	const token = localStorage.getItem('token');
-	if (token) {
-		const response = await getUserRole(token);
-		if (response && response.ok) {
-			return await response.text();
-		} else {
-			//throw new Error("Couldn't access user's role.");
-			console.error("Couldn't access user's role.");
-		}
-	} else {
-		//throw new Error('You are not authorized!');
-		console.error('You are not authorized!');
 	}
 };
